requirements.txt
streamlit
google-generativeai
pandas
os